+++ CHANGES.rst
@@ -37,6 +37,7 @@ Bugfixes
 - Fix error when sending scheduled reminders using an event time placeholder (:pr:`7238`)
 - Fix rendering static text/images in poster templates (:pr:`7239`)
 - Do not revert to defaults when disabling all optional event cloners (:issue:`6833`, :pr:`7245`)
+- Fix changing tracks for invited abstracts (:issue:`7061`, :pr:`7240`)
 
 Accessibility
 ^^^^^^^^^^^^^